feat: add payment to storage module (#44)
* add payment to storage module

rename function
add update bucket info message
add source type check
init payment in storage module

* format

* fix test and lint

* fixed golangci lint

* rename price_time and out_flows

* add charge

* Add a todo to ReadQuota

* fix review tips

* fix ci lint

* fix OutFlowInUSD and add some todo

* leave a todo on setBucket/Object

+++ x/storage/keeper/msg_server.go
@@ -6,6 +6,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
+	paymenttypes "github.com/bnb-chain/greenfield/x/payment/types"
 	sptypes "github.com/bnb-chain/greenfield/x/sp/types"
 	"github.com/bnb-chain/greenfield/x/storage/types"
 )
@@ -40,6 +41,9 @@ func (k msgServer) CreateBucket(goCtx context.Context, msg *types.MsgCreateBucke
 		if err != nil {
 			return nil, err
 		}
+		if !k.paymentKeeper.IsPaymentAccountOwner(ctx, paymentAcc.String(), ownerAcc.String()) {
+			return nil, paymenttypes.ErrNotPaymentAccountOwner
+		}
 	} else {
 		paymentAcc = ownerAcc
 	}
@@ -48,7 +52,6 @@ func (k msgServer) CreateBucket(goCtx context.Context, msg *types.MsgCreateBucke
 	if err != nil {
 		return nil, err
 	}
-
 	// TODO: this is a very tricky implement. Will be refactor later.
 	spApproval := msg.PrimarySpApprovalSignature
 	msg.PrimarySpApprovalSignature = []byte("")
@@ -57,7 +60,7 @@ func (k msgServer) CreateBucket(goCtx context.Context, msg *types.MsgCreateBucke
 		return nil, err
 	}
 
-	err = k.CheckSPAndSignature(ctx, []string{msg.PrimarySpAddress}, [][]byte{sdk.Keccak256(bz)}, [][]byte{spApproval})
+	err = k.VerifySPAndSignature(ctx, []string{msg.PrimarySpAddress}, [][]byte{sdk.Keccak256(bz)}, [][]byte{spApproval})
 	if err != nil {
 		return nil, err
 	}
@@ -69,9 +72,17 @@ func (k msgServer) CreateBucket(goCtx context.Context, msg *types.MsgCreateBucke
 		IsPublic:         msg.IsPublic,
 		CreateAt:         ctx.BlockHeight(),
 		Id:               k.GetBucketId(ctx),
+		SourceType:       types.SOURCE_TYPE_ORIGIN,
 		ReadQuota:        types.READ_QUOTA_FREE,
 		PaymentAddress:   paymentAcc.String(),
 		PrimarySpAddress: primarySPAcc.String(),
+	}
+
+	if msg.ReadQuota != types.READ_QUOTA_FREE {
+		err := k.paymentKeeper.ChargeInitialReadFee(ctx, &bucketInfo)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	return &types.MsgCreateBucketResponse{}, k.Keeper.CreateBucket(ctx, bucketInfo)
@@ -80,8 +91,49 @@ func (k msgServer) CreateBucket(goCtx context.Context, msg *types.MsgCreateBucke
 func (k msgServer) DeleteBucket(goCtx context.Context, msg *types.MsgDeleteBucket) (*types.MsgDeleteBucketResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
+	bucketInfo, found := k.Keeper.GetBucket(ctx, msg.BucketName)
+	if !found {
+		return nil, types.ErrNoSuchBucket
+	}
+	if bucketInfo.SourceType != types.SOURCE_TYPE_ORIGIN {
+		return nil, types.ErrSourceTypeMismatch
+	}
 	// TODO: check if have the permission to delete bucket
 	return &types.MsgDeleteBucketResponse{}, k.Keeper.DeleteBucket(ctx, msg.BucketName)
+}
+
+func (k msgServer) UpdateBucketInfo(goCtx context.Context, msg *types.MsgUpdateBucketInfo) (*types.MsgUpdateBucketInfoResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	bucketInfo, found := k.Keeper.GetBucket(ctx, msg.BucketName)
+	if !found {
+		return nil, types.ErrNoSuchBucket
+	}
+	if bucketInfo.SourceType != types.SOURCE_TYPE_ORIGIN {
+		return nil, types.ErrSourceTypeMismatch
+	}
+
+	if msg.PaymentAddress != "" && msg.PaymentAddress != bucketInfo.PaymentAddress {
+		if !k.paymentKeeper.IsPaymentAccountOwner(ctx, bucketInfo.Owner, msg.PaymentAddress) {
+			return nil, paymenttypes.ErrNotPaymentAccountOwner
+		}
+		err := k.paymentKeeper.ChargeUpdatePaymentAccount(ctx, &bucketInfo, &msg.PaymentAddress)
+		if err != nil {
+			return nil, err
+		}
+		bucketInfo.PaymentAddress = msg.PaymentAddress
+	}
+
+	if msg.ReadQuota != bucketInfo.ReadQuota {
+		err := k.paymentKeeper.ChargeUpdateReadQuota(ctx, &bucketInfo, msg.ReadQuota)
+		if err != nil {
+			return nil, err
+		}
+		bucketInfo.ReadQuota = msg.ReadQuota
+	}
+
+	k.Keeper.SetBucket(ctx, bucketInfo)
+	return &types.MsgUpdateBucketInfoResponse{}, nil
 }
 
 func (k msgServer) CreateObject(goCtx context.Context, msg *types.MsgCreateObject) (*types.MsgCreateObjectResponse, error) {
@@ -108,26 +160,29 @@ func (k msgServer) CreateObject(goCtx context.Context, msg *types.MsgCreateObjec
 	if err != nil {
 		return nil, err
 	}
-	err = k.CheckSPAndSignature(ctx, []string{bucketInfo.PrimarySpAddress}, [][]byte{sdk.Keccak256(bz)}, [][]byte{spApproval})
+	err = k.VerifySPAndSignature(ctx, []string{bucketInfo.PrimarySpAddress}, [][]byte{sdk.Keccak256(bz)}, [][]byte{spApproval})
 	if err != nil {
 		return nil, err
 	}
 
 	objectInfo := types.ObjectInfo{
-		Owner:          ownerAcc.String(),
-		BucketName:     msg.BucketName,
-		ObjectName:     msg.ObjectName,
-		PayloadSize:    msg.PayloadSize,
-		IsPublic:       msg.IsPublic,
-		ContentType:    msg.ContentType,
-		Id:             k.GetObjectID(ctx),
-		CreateAt:       ctx.BlockHeight(),
-		ObjectStatus:   types.OBJECT_STATUS_INIT,
-		RedundancyType: types.REDUNDANCY_EC_TYPE, // TODO: base on redundancy policy
-		SourceType:     types.SOURCE_TYPE_ORIGIN,
-
+		Owner:                ownerAcc.String(),
+		BucketName:           msg.BucketName,
+		ObjectName:           msg.ObjectName,
+		PayloadSize:          msg.PayloadSize,
+		IsPublic:             msg.IsPublic,
+		ContentType:          msg.ContentType,
+		Id:                   k.GetObjectID(ctx),
+		CreateAt:             ctx.BlockHeight(),
+		ObjectStatus:         types.OBJECT_STATUS_INIT,
+		RedundancyType:       types.REDUNDANCY_EC_TYPE, // TODO: base on redundancy policy
+		SourceType:           types.SOURCE_TYPE_ORIGIN,
 		Checksums:            msg.ExpectChecksums,
 		SecondarySpAddresses: msg.ExpectSecondarySpAddresses,
+	}
+	err = k.paymentKeeper.LockStoreFee(ctx, &bucketInfo, &objectInfo)
+	if err != nil {
+		return nil, err
 	}
 
 	return &types.MsgCreateObjectResponse{}, k.Keeper.CreateObject(ctx, objectInfo)
@@ -142,7 +197,7 @@ func (k msgServer) SealObject(goCtx context.Context, msg *types.MsgSealObject) (
 	if err != nil {
 		return nil, err
 	}
-	bucketInfo, found := k.Keeper.GetBucket(ctx, msg.BucketName)
+	bucketInfo, found := k.GetBucket(ctx, msg.BucketName)
 	if !found {
 		return nil, types.ErrNoSuchBucket
 	}
@@ -150,8 +205,7 @@ func (k msgServer) SealObject(goCtx context.Context, msg *types.MsgSealObject) (
 	if bucketInfo.PrimarySpAddress != spAcc.String() {
 		return nil, types.ErrSPAddressMismatch
 	}
-
-	objectInfo, found := k.Keeper.GetObject(ctx, msg.BucketName, msg.ObjectName)
+	objectInfo, found := k.GetObject(ctx, msg.BucketName, msg.ObjectName)
 	if !found {
 		return nil, types.ErrNoSuchObject
 	}
@@ -161,12 +215,21 @@ func (k msgServer) SealObject(goCtx context.Context, msg *types.MsgSealObject) (
 		objectInfo.ObjectStatus = types.OBJECT_STATUS_IN_SERVICE
 	}
 
-	err = k.CheckSPAndSignature(ctx, msg.SecondarySpAddresses, objectInfo.Checksums[1:], msg.SecondarySpSignatures)
+	err = k.VerifySPAndSignature(ctx, msg.SecondarySpAddresses, objectInfo.Checksums[1:], msg.SecondarySpSignatures)
 	if err != nil {
 		return nil, err
 	}
 
-	k.Keeper.SetObject(ctx, objectInfo)
+	err = k.paymentKeeper.UnlockAndChargeStoreFee(ctx, &bucketInfo, &objectInfo)
+	if err != nil {
+		return nil, err
+	}
+	objectInfo.LockedBalance = nil
+
+	// TODO(fynn): SetBucket/Object will cost a lot every time we set it. So maybe need to split the info to two types
+	// key-value, one is immutable attributes, another is mutable attributes, to reduce performance overhead
+	k.SetBucket(ctx, bucketInfo)
+	k.SetObject(ctx, objectInfo)
 
 	return &types.MsgSealObjectResponse{}, nil
 }
@@ -194,6 +257,10 @@ func (k msgServer) CopyObject(goCtx context.Context, msg *types.MsgCopyObject) (
 		return nil, sdkerrors.Wrapf(types.ErrNoSuchObject, "src object name (%s)", msg.SrcObjectName)
 	}
 
+	if srcObjectInfo.SourceType != types.SOURCE_TYPE_ORIGIN {
+		return nil, types.ErrSourceTypeMismatch
+	}
+
 	// TODO: this is a very tricky implement. Will be refactor later.
 	spApproval := msg.DstPrimarySpApprovalSignature
 	msg.DstPrimarySpApprovalSignature = []byte("")
@@ -201,7 +268,7 @@ func (k msgServer) CopyObject(goCtx context.Context, msg *types.MsgCopyObject) (
 	if err != nil {
 		return nil, err
 	}
-	err = k.CheckSPAndSignature(ctx, []string{dstBucketInfo.PrimarySpAddress}, [][]byte{sdk.Keccak256(bz)}, [][]byte{spApproval})
+	err = k.VerifySPAndSignature(ctx, []string{dstBucketInfo.PrimarySpAddress}, [][]byte{sdk.Keccak256(bz)}, [][]byte{spApproval})
 	if err != nil {
 		return nil, err
 	}
@@ -222,10 +289,13 @@ func (k msgServer) CopyObject(goCtx context.Context, msg *types.MsgCopyObject) (
 		ObjectStatus:   types.OBJECT_STATUS_INIT,
 		RedundancyType: types.REDUNDANCY_EC_TYPE,
 		SourceType:     types.SOURCE_TYPE_ORIGIN,
-
-		Checksums: srcObjectInfo.Checksums,
+		Checksums:      srcObjectInfo.Checksums,
 	}
 
+	err = k.paymentKeeper.LockStoreFee(ctx, &dstBucketInfo, &objectInfo)
+	if err != nil {
+		return nil, err
+	}
 	k.Keeper.SetObject(ctx, objectInfo)
 
 	return &types.MsgCopyObjectResponse{}, nil
@@ -234,9 +304,22 @@ func (k msgServer) CopyObject(goCtx context.Context, msg *types.MsgCopyObject) (
 func (k msgServer) DeleteObject(goCtx context.Context, msg *types.MsgDeleteObject) (*types.MsgDeleteObjectResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
+	bucketInfo, found := k.Keeper.GetBucket(ctx, msg.BucketName)
+	if !found {
+		return nil, types.ErrNoSuchBucket
+	}
+
 	objectInfo, found := k.Keeper.GetObject(ctx, msg.BucketName, msg.ObjectName)
 	if !found {
 		return nil, types.ErrNoSuchObject
+	}
+
+	if objectInfo.SourceType != types.SOURCE_TYPE_ORIGIN {
+		return nil, types.ErrSourceTypeMismatch
+	}
+
+	if objectInfo.ObjectStatus != types.OBJECT_STATUS_IN_SERVICE {
+		return nil, types.ErrObjectNotInService
 	}
 
 	// Currently, only the owner is allowed to delete object
@@ -244,6 +327,10 @@ func (k msgServer) DeleteObject(goCtx context.Context, msg *types.MsgDeleteObjec
 		return nil, types.ErrAccessDenied
 	}
 
+	err := k.paymentKeeper.ChargeDeleteObject(ctx, &bucketInfo, &objectInfo)
+	if err != nil {
+		return nil, err
+	}
 	k.Keeper.DeleteObject(ctx, msg.BucketName, msg.ObjectName)
 	return &types.MsgDeleteObjectResponse{}, nil
 }
@@ -251,13 +338,17 @@ func (k msgServer) DeleteObject(goCtx context.Context, msg *types.MsgDeleteObjec
 func (k msgServer) RejectSealObject(goCtx context.Context, msg *types.MsgRejectSealObject) (*types.MsgRejectSealObjectResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
+	bucketInfo, found := k.Keeper.GetBucket(ctx, msg.BucketName)
+	if !found {
+		return nil, types.ErrNoSuchBucket
+	}
 	objectInfo, found := k.Keeper.GetObject(ctx, msg.BucketName, msg.ObjectName)
 	if !found {
 		return nil, types.ErrNoSuchObject
 	}
 
 	if objectInfo.ObjectStatus != types.OBJECT_STATUS_INIT {
-		return nil, types.ErrObjectStatusNotInit
+		return nil, types.ErrObjectNotInit
 	}
 
 	// Currently, only the primary sp is allowed to reject seal object
@@ -273,8 +364,11 @@ func (k msgServer) RejectSealObject(goCtx context.Context, msg *types.MsgRejectS
 	if sp.Status != sptypes.STATUS_IN_SERVICE {
 		return nil, types.ErrStorageProviderNotInService
 	}
+	err = k.paymentKeeper.UnlockStoreFee(ctx, &bucketInfo, &objectInfo)
+	if err != nil {
+		return nil, err
+	}
 
-	// TODO: Interact with payment. unlock the pre-pay fee.
 	k.Keeper.DeleteObject(ctx, msg.BucketName, msg.ObjectName)
 	return &types.MsgRejectSealObjectResponse{}, nil
 }
@@ -283,9 +377,10 @@ func (k msgServer) CreateGroup(goCtx context.Context, msg *types.MsgCreateGroup)
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
 	groupInfo := types.GroupInfo{
-		Owner:     msg.Creator,
-		Id:        k.GetGroupId(ctx),
-		GroupName: msg.GroupName,
+		Owner:      msg.Creator,
+		SourceType: types.SOURCE_TYPE_ORIGIN,
+		Id:         k.GetGroupId(ctx),
+		GroupName:  msg.GroupName,
 	}
 	err := k.Keeper.CreateGroup(ctx, groupInfo)
 	if err != nil {
@@ -314,6 +409,13 @@ func (k msgServer) CreateGroup(goCtx context.Context, msg *types.MsgCreateGroup)
 func (k msgServer) DeleteGroup(goCtx context.Context, msg *types.MsgDeleteGroup) (*types.MsgDeleteGroupResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
+	groupInfo, found := k.Keeper.GetGroup(ctx, msg.Operator, msg.GroupName)
+	if !found {
+		return nil, types.ErrNoSuchGroup
+	}
+	if groupInfo.SourceType != types.SOURCE_TYPE_ORIGIN {
+		return nil, types.ErrSourceTypeMismatch
+	}
 	// Note: Delete group does not require the group is empty. The group member will be deleted by on-chain GC.
 	return &types.MsgDeleteGroupResponse{}, k.Keeper.DeleteGroup(ctx, msg.Operator, msg.GroupName)
 }
@@ -325,6 +427,9 @@ func (k msgServer) LeaveGroup(goCtx context.Context, msg *types.MsgLeaveGroup) (
 	if !found {
 		return nil, types.ErrNoSuchGroup
 	}
+	if groupInfo.SourceType != types.SOURCE_TYPE_ORIGIN {
+		return nil, types.ErrSourceTypeMismatch
+	}
 	return &types.MsgLeaveGroupResponse{}, k.Keeper.RemoveGroupMember(ctx, groupInfo.Id, msg.Member)
 }
 
@@ -335,6 +440,9 @@ func (k msgServer) UpdateGroupMember(goCtx context.Context, msg *types.MsgUpdate
 	groupInfo, found := k.Keeper.GetGroup(ctx, msg.Operator, msg.GroupName)
 	if !found {
 		return nil, types.ErrNoSuchGroup
+	}
+	if groupInfo.SourceType != types.SOURCE_TYPE_ORIGIN {
+		return nil, types.ErrSourceTypeMismatch
 	}
 
 	for _, member := range msg.MembersToAdd {
